requirements-clean.txt
aiohttp==3.12.14
appnope==0.1.4
argcomplete==3.6.2
argon2_cffi_bindings==21.2.0
asyncssh_unofficial==0.9.2
atheris==2.3.0
BeaEnginePython==3.1.0
brotli==1.1.0
brotlicffi==1.1.0.0
cached_property==2.0.1
capstone==5.0.6
cchardet==2.1.7
cerealizer==0.8.4
cffi==1.17.1
cloudpickle==3.1.1
clr==1.0.3
colorama==0.4.6
ConfigParser==7.2.0
copier==9.8.0
coverage==7.9.2
curio==1.6
Cython==3.1.2
distorm3==3.5.2
dl==0.1.0
docrepr==0.2.0
flake8==7.3.0
Foundation==0.1.0a0.dev1
gevent==25.5.1
gobject==0.1.0
greenlet==3.2.3
h2==4.2.0
HTMLParser==0.0.2
hypothesis==6.135.33
importlib_resources==6.5.2
ini2toml==0.15
ipyparallel==9.0.1
ipywidgets==8.1.7
jinxed==1.3.0
jnius==1.1.0
js==1.0
jupyter_ai==2.31.5
jupyter_ai_magics==2.31.5
jupyter_collaboration==4.1.0
jupyterhub==5.3.0
keyring==25.6.0
matplotlib==3.10.3
mtrand==0.1
multidict==6.6.3
mypy==1.17.0
netifaces==0.11.0
nose==1.3.7
numpydoc==1.9.0
openapi_core==0.19.5
outcome==1.3.0.post0
pandas==2.3.1
pathlib2==2.3.7.post1
Pillow==11.3.0
pip_api==0.0.34
pkg1==0.0.3
pkgutil_resolve_name==1.3.10
playwright==1.53.0
pox==0.3.6
psyco==1.6
pycares==4.9.0
pycurl==7.45.6
pyczmq==0.0.4
pyenchant==3.2.2
PyInstaller==6.14.2
pylama==8.4.1
pyodide==0.0.2
PyOpenGL==3.1.9
pyparsing==3.2.3
pyperf==2.9.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.9.1
PySide==1.2.4
Pyside==1.2.4
PySide2==5.15.2.1
PySide6==6.9.1
pysqlite==2.8.3
python_socks==2.7.1
regex==2024.11.6
rfc3987==1.3.8
ruamel.base==1.0.0
scipy_doctest==1.8.0
sip==6.12.0
socksio==1.0.0
Sphinx==8.2.3
SQLAlchemy==2.0.41
stackless==0.0.1
sympy==1.14.0
testpath==0.6.0
threadpoolctl==3.6.0
tokenize_rt==6.2.0
trio_typing==0.10.0
trove_classifiers==2025.5.9.12
urllib3_secure_extra==0.1.0
uvloop==0.21.0
websockets==15.0.1
win32security==2.1.0
wmi==1.5.1
wsaccel==0.6.7
xmlrpclib==1.0.1
xx==3.3.2
yapf==0.43.0
zstandard==0.23.0
